requirements.txt
Pillow
django
requests
